Add EEPROM programming script
Due to changes in wifiService app behaviour, some test runs failed for some
early production runs.  These units were later manually tested and found to
be working. But, because the factory test failed, those units did not have
their EEPROMs programmed. This script was created for such cases, to allow
those units to have their EEPROMs programmed in less time than it would take
to run them through the entire factory programming/test process again.
This is NOT intended to be used on fresh units because it skips the
.spk image programming step and doesn't run any tests.